AN-4: moved setupInitialData to pleo-antaeus-data module for further reuse in the integration tests
